Release fwupd 1.8.9

+++ src/fu-daemon.c
@@ -234,7 +234,7 @@ fu_daemon_create_request(FuDaemon *self, const gchar *sender, GError **error)
 {
 	FuDaemonSenderItem *sender_item;
 	FwupdDeviceFlags device_flags = FWUPD_DEVICE_FLAG_NONE;
-	guint64 calling_uid = 0;
+	guint calling_uid = 0;
 	g_autoptr(FuEngineRequest) request = fu_engine_request_new(FU_ENGINE_REQUEST_KIND_ACTIVE);
 	g_autoptr(GVariant) value = NULL;
 
@@ -843,7 +843,7 @@ fu_daemon_install_with_helper_device(FuMainAuthHelper *helper,
 #else
 		rels = xb_node_query(component, "releases/release", 0, NULL);
 #endif
-		/* add all but the the first entry */
+		/* add all but the first entry */
 		for (guint i = 1; i < rels->len; i++) {
 			XbNode *rel = g_ptr_array_index(rels, i);
 			g_autoptr(FuRelease) release2 = fu_release_new();
